see desc!
API:
Added NachoPlayer#sendActionBar
Added NachoPlayer#jump
Added Unsafe#sendPacket
Fixed /version command

Server:
Removed a lot of useless code and casts
changed Server#getName to the proper brand name

+++ NachoSpigot-Server/src/main/java/net/minecraft/server/EntityHuman.java
@@ -94,10 +94,10 @@ public abstract class EntityHuman extends EntityLiving {
 
     protected void h() {
         super.h();
-        this.datawatcher.a(16, Byte.valueOf((byte) 0));
-        this.datawatcher.a(17, Float.valueOf(0.0F));
-        this.datawatcher.a(18, Integer.valueOf(0));
-        this.datawatcher.a(10, Byte.valueOf((byte) 0));
+        this.datawatcher.a(16, (byte) 0);
+        this.datawatcher.a(17, 0.0F);
+        this.datawatcher.a(18, 0);
+        this.datawatcher.a(10, (byte) 0);
     }
 
     public boolean bS() {
